Improve authentication process (#2)
Keep the refresh token in a custom credentials storage. Thanks to this, whenever the refresh token changes (which happens from time to time), it's remembered in the persistent storage, so it can sustain Node-RED restarts.

Force user to save and deploy the configuration node before triggering OAuth flow. Thanks to this, the secret is safely stored inside Node-RED configuration, and it's a less chance for it to leak.

The OAuth's redirect page has been slightly improved, so it doesn't look like a scam (that much).

Committed changes
-----------------

* Small improvements.

* Idea project configuration.

* Implement a helper lib for storing encrypted credentials.

This is required in order to periodically store new refresh tokens. Sadly, Node-RED doesn't allow to programmatically update credentials, so it must be done on our own.

* Add debugging helpers to Dockerfile.

* Store refresh token in the custom credentials storage.

* Store refresh token in the custom credentials storage.

* Read the refresh token from the custom credentials store.

When the node has been removed, also remove the file with credentials.

* Make the redirect page prettier, and move it to a separate file.

* Add hidden form field to track authentication progress.

The form field is required, and is set when the OAuth flow completes. This way, the configuration node can be shown as invalid until the user finishes the whole OAuth flow.